+++ lib/kithe/indexable_settings.rb
@@ -1,14 +1,17 @@
 module Kithe
   class IndexableSettings
     attr_accessor :solr_url, :writer_class_name, :writer_settings,
-                  :model_name_solr_field, :solr_id_value_attribute, :disable_callbacks
+                  :model_name_solr_field, :solr_id_value_attribute, :disable_callbacks,
+                  :batching_mode_batch_size
     def initialize(solr_url:, writer_class_name:, writer_settings:,
-                   model_name_solr_field:, solr_id_value_attribute:, disable_callbacks: false)
+                   model_name_solr_field:, solr_id_value_attribute:, disable_callbacks: false,
+                   batching_mode_batch_size: 100)
       @solr_url = solr_url
       @writer_class_name = writer_class_name
       @writer_settings = writer_settings
       @model_name_solr_field = model_name_solr_field
       @solr_id_value_attribute = solr_id_value_attribute || 'id'
+      @batching_mode_batch_size = batching_mode_batch_size
     end
 
     # Use configured solr_url, and merge together with configured